metadata for fields
